Move fixtures to their own file
