+++ dart_dependency_validator.yaml
@@ -1,0 +1,2 @@
+exclude:
+  - test/tools/fixtures/sort_imports/imports.dart
